cleanup: move naivepath back to pathingentity
